moves _ObservableDict tests into separate file
