+++ tests/observer_pattern/observable/test_observable_object.py
@@ -69,66 +69,6 @@ def test_class_object_list_attribute(caplog: pytest.LogCaptureFixture) -> None:
     assert "'list_attr[0].name' changed to 'Ciao'" in caplog.text
 
 
-def test_simple_instance_dict_attribute(caplog: pytest.LogCaptureFixture) -> None:
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.dict_attr = {"first": "Hello"}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.dict_attr["first"] = "Ciao"
-    instance.dict_attr["second"] = "World"
-
-    assert "'dict_attr[\"first\"]' changed to 'Ciao'" in caplog.text
-    assert "'dict_attr[\"second\"]' changed to 'World'" in caplog.text
-
-
-def test_simple_class_dict_attribute(caplog: pytest.LogCaptureFixture) -> None:
-    class MyObservable(Observable):
-        dict_attr = {"first": "Hello"}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.dict_attr["first"] = "Ciao"
-    instance.dict_attr["second"] = "World"
-
-    assert "'dict_attr[\"first\"]' changed to 'Ciao'" in caplog.text
-    assert "'dict_attr[\"second\"]' changed to 'World'" in caplog.text
-
-
-def test_instance_dict_attribute(caplog: pytest.LogCaptureFixture) -> None:
-    class NestedObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.name = "Hello"
-
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.dict_attr = {"first": NestedObservable()}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.dict_attr["first"].name = "Ciao"
-
-    assert "'dict_attr[\"first\"].name' changed to 'Ciao'" in caplog.text
-
-
-def test_class_dict_attribute(caplog: pytest.LogCaptureFixture) -> None:
-    class NestedObservable(Observable):
-        name = "Hello"
-
-    class MyObservable(Observable):
-        dict_attr = {"first": NestedObservable()}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.dict_attr["first"].name = "Ciao"
-
-    assert "'dict_attr[\"first\"].name' changed to 'Ciao'" in caplog.text
-
-
 def test_removed_observer_on_class_list_attr(caplog: pytest.LogCaptureFixture) -> None:
     class NestedObservable(Observable):
         name = "Hello"
@@ -150,44 +90,6 @@ def test_removed_observer_on_class_list_attr(caplog: pytest.LogCaptureFixture) -
 
     assert "'nested_attr.name' changed to 'Hi'" in caplog.text
     assert "'changed_list_attr[0].name' changed to 'Hi'" not in caplog.text
-
-
-def test_removed_observer_on_instance_dict_attr(
-    caplog: pytest.LogCaptureFixture,
-) -> None:
-    class NestedObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.name = "Hello"
-
-    nested_instance = NestedObservable()
-
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.nested_attr = nested_instance
-            self.changed_dict_attr = {"nested": nested_instance, 2.1: nested_instance}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.changed_dict_attr["nested"] = "Ciao"
-    instance.changed_dict_attr[2.1] = "foo"
-
-    assert "'changed_dict_attr[\"nested\"]' changed to 'Ciao'" in caplog.text
-    assert "'changed_dict_attr[2.1]' changed to 'foo'" in caplog.text
-    caplog.clear()
-
-    assert nested_instance._observers == {
-        '["nested"]': [],
-        "[2.1]": [],
-        "nested_attr": [instance],
-    }
-
-    instance.nested_attr.name = "Hi"
-
-    assert "'nested_attr.name' changed to 'Hi'" in caplog.text
-    assert "'changed_dict_attr[\"nested\"].name' changed to 'Hi'" not in caplog.text
-    assert "'changed_dict_attr[2.1].name' changed to 'Hi'" not in caplog.text
 
 
 def test_removed_observer_on_instance_list_attr(
@@ -217,86 +119,6 @@ def test_removed_observer_on_instance_list_attr(
 
     assert "'nested_attr.name' changed to 'Hi'" in caplog.text
     assert "'changed_list_attr[0].name' changed to 'Hi'" not in caplog.text
-
-
-def test_removed_observer_on_class_dict_attr(caplog: pytest.LogCaptureFixture) -> None:
-    class NestedObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.name = "Hello"
-
-    nested_instance = NestedObservable()
-    assert nested_instance._observers == {}
-
-    class MyObservable(Observable):
-        nested_attr = nested_instance
-        changed_dict_attr = {"nested": nested_instance, 2.1: nested_instance}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.changed_dict_attr["nested"] = "Ciao"
-    instance.changed_dict_attr[2.1] = "foo"
-
-    assert "'changed_dict_attr[\"nested\"]' changed to 'Ciao'" in caplog.text
-    assert "'changed_dict_attr[2.1]' changed to 'foo'" in caplog.text
-    caplog.clear()
-
-    assert nested_instance._observers == {
-        '["nested"]': [],
-        "[2.1]": [],
-        "nested_attr": [instance],
-    }
-
-    instance.nested_attr.name = "Hi"
-
-    assert "'nested_attr.name' changed to 'Hi'" in caplog.text
-    assert "'changed_dict_attr[\"nested\"].name' changed to 'Hi'" not in caplog.text
-    assert "'changed_dict_attr[2.1].name' changed to 'Hi'" not in caplog.text
-
-
-def test_nested_dict_instances(caplog: pytest.LogCaptureFixture) -> None:
-    dict_instance = {"first": "Hello", "second": "World"}
-
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.nested_dict_attr = {"nested": dict_instance}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.nested_dict_attr["nested"]["first"] = "Ciao"
-
-    assert "'nested_dict_attr[\"nested\"][\"first\"]' changed to 'Ciao'" in caplog.text
-
-
-def test_dict_in_list_instance(caplog: pytest.LogCaptureFixture) -> None:
-    dict_instance = {"first": "Hello", "second": "World"}
-
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.dict_in_list = [dict_instance]
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.dict_in_list[0]["first"] = "Ciao"
-
-    assert "'dict_in_list[0][\"first\"]' changed to 'Ciao'" in caplog.text
-
-
-def test_list_in_dict_instance(caplog: pytest.LogCaptureFixture) -> None:
-    list_instance: list[Any] = [1, 2, 3]
-
-    class MyObservable(Observable):
-        def __init__(self) -> None:
-            super().__init__()
-            self.list_in_dict = {"some_list": list_instance}
-
-    instance = MyObservable()
-    MyObserver(instance)
-    instance.list_in_dict["some_list"][0] = "Ciao"
-
-    assert "'list_in_dict[\"some_list\"][0]' changed to 'Ciao'" in caplog.text
 
 
 def test_list_append(caplog: pytest.LogCaptureFixture) -> None:
